Add Websockets for bid history

+++ RzhadBids/DTO/LotDTO.cs
@@ -7,7 +7,7 @@ namespace RzhadBids.DTO
         public string Title { get; set; }
         public string? Description { get; set; }
         public int CategoryId { get; set; }
-        public decimal StartingPrice { get; set; }
+        public int StartingPrice { get; set; }
         public DateTime DateStart { get; set; }
         public DateTime DateEnd { get; set; }
         public int? ChatId { get; set; }
